Rename to `angular-ignorer`
