+++ packages/instrumenter/src/frameworks/angular-ignorer.ts
