+++ packages/instrumenter/src/frameworks/index.ts
@@ -1,6 +1,6 @@
 import { PluginKind, declareClassPlugin } from '@stryker-mutator/api/plugin';
 
-import { AngularIgnorer } from './angular-ignore.js';
+import { AngularIgnorer } from './angular-ignorer.js';
 
 export const strykerPlugins = [declareClassPlugin(PluginKind.Ignore, 'angular', AngularIgnorer)];
